fix image path issue
