+++ .theme/components/Prose/ProseImg.vue
@@ -1,17 +1,14 @@
 <template>
-  <img
-    :src="refinedSrc"
+  <Image
+    :src="src"
     :alt="alt"
     :width="width"
     :height="height"
-    :class="[ alt ]"
-  >
+    :class="alt"
+  />
 </template>
-<script setup lang="ts">
-import { withTrailingSlash, withLeadingSlash, joinURL } from "ufo";
-import { useRuntimeConfig, computed, useRoute } from "#imports";
-
-const props = defineProps({
+<script setup>
+defineProps({
   src: {
     type: String,
     default: ""
@@ -29,25 +26,6 @@ const props = defineProps({
     default: undefined
   }
 });
-
-const route = useRoute();
-
-const refinedSrc = computed(() => {
-  if(props.src.startsWith("//")) {
-    return props.src;
-  }
-
-  const base = withLeadingSlash(withTrailingSlash(useRuntimeConfig().app.baseURL));
-  const path = parsePath(route.path);
-
-  return joinURL(base, path, props.src);
-});
-
-function parsePath(path: string) {
-  return path.endsWith("/")
-    ? path
-    : path.replace(/\/[^/]*\/?$/, "");
-}
 </script>
 <style lang="scss" scoped>
 img {
